requirements.txt
setuptools~=58.0.0
pandas~=1.3.2
numpy~=1.21.2
openpyxl~=3.0.7
parse~=1.19.0
simplejson~=3.17.5
dash~=2.0.0
plotly~=5.3.1
dash-daq~=0.5.0
python-igraph~=0.9.6
matplotlib~=3.4.3